requirements.txt
absl-py==1.4.0
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==4.0.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
geotorch==0.3.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.12.0
idna==2.10
imageio==2.25.1
importlib-metadata==6.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
pandas==1.3.5
pathlib2==2.3.7.post1
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.1
robustbench @ git+https://github.com/RobustBench/robustbench.git@0d6a12e2e04b12d74e508701a9201cdc445d9b0d
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.2
sewar==0.4.5
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
timm==0.6.12
torch==1.13.1
torchdiffeq==0.2.3
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
pyrtools==1.0.1
